+++ src/main/java/com/glory/app/rest/Repo/UserRepo.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepo extends JpaRepository<User, Long> {
+public interface UserRepo extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
+
 }
